🚚 Restructure bookstore routes
