+++ src/pages/store/books/index.vue
@@ -1,0 +1,7 @@
+<script>
+export default {
+  fetch({ redirect, localeLocation }) {
+    redirect(localeLocation({ name: 'store-books-featured' }));
+  },
+};
+</script>
